analyse/requirements.txt
kafka-python==2.0.2
influxdb-client==1.46.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
